%% main
\name{OmicCircos-package}
\alias{OmicCircos}
\title{
OmicCircos package
}
\description{
 OmicCircos is for generating high-quality circular plots for omics data.
}
\author{
%%  ~~ possibly secondary sources and usages ~~
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}
\references{
%%  ~~ possibly secondary sources and usages ~~
 OmicCircos: an R package for simple and circular visualization of omics data. Cancer Inform. 2014 Jan 16;13:13-20. doi: 10.4137/CIN.S13495. eCollection 2014. PMID: 24526832 [PubMed] PMCID: PMC3921174
}
%%\keyword{key}
