%%
\name{TCGA.BC.cnv.2k.60}
\alias{TCGA.BC.cnv.2k.60}
\docType{data}
\title{
 copy number data of TCGA Breast Cancer 
}
\description{
 Examples of TCGA Breast Cancer DNA copy number data from 500 genes and 60 samples.
}
\author{
%%  ~~ possibly secondary sources and usages ~~
Ying Hu <yhu@mail.nih.gov>
Chunhua Yan <yanch@mail.nih.gov>
}



