%%
\name{TCGA.BC.fus}
\alias{TCGA.BC.fus}
\docType{data}
\title{
 TCGA Breast Cancer gene fusion data.
}
\description{
 Examples of TCGA Breast Cancer gene fusion data from 18 fusion proteins
}
\author{
%%  ~~ possibly secondary sources and usages ~~
Ying Hu <yhu@mail.nih.gov>
Chunhua Yan <yanch@mail.nih.gov>
}



