%%
\name{TCGA.BC.gene.exp.2k.60}
\alias{TCGA.BC.gene.exp.2k.60}
\docType{data}
\title{
 TCGA BRCA expression data
}
\description{
 Examples of TCGA Breast Cancer expression data from 500 genes and 60 samples
}
\author{
%%  ~~ possibly secondary sources and usages ~~
Ying Hu <yhu@mail.nih.gov>
Chunhua Yan <yanch@mail.nih.gov>
}



