%%
\name{TCGA.BC.sample60}
\alias{TCGA.BC.sample60}
\docType{data}
\title{
 TCGA BRCA Sample names and subtypes
}
\description{
 Names and subtypes of 60 samples for TCGA Breast Cancer expression and copy number data 
}
\author{
%%  ~~ possibly secondary sources and usages ~~
Ying Hu <yhu@mail.nih.gov>
Chunhua Yan <yanch@mail.nih.gov>
}



