%%
\name{TCGA.BC_Her2_cnv_exp}
\alias{TCGA.BC_Her2_cnv_exp}
\docType{data}
\title{
 TCGA BRCA expression and cnv association
}
\description{
 The p-values of associations between the TCGA Breast Cancer copy number and gene expression data.
}
\author{
%%  ~~ possibly secondary sources and usages ~~
Ying Hu <yhu@mail.nih.gov>
Chunhua Yan <yanch@mail.nih.gov>
}



