%%
\name{UCSC.hg18.chr}
\alias{UCSC.hg18.chr}
\docType{data}
\title{
 human hg18 segment data.
}
\description{
 Human hg18 chromome size and binding data obtained from UCSC Genome Browser cytogenetics table.
}
\author{
%%  ~~ possibly secondary sources and usages ~~
Ying Hu <yhu@mail.nih.gov>
Chunhua Yan <yanch@mail.nih.gov>
}



