%%
\name{UCSC.hg18}
\alias{UCSC.hg18}
\docType{data}
\title{
 human hg18 circumference coordinates
}
\description{
 Human hg18 circumference coordinates (angles) are calculated from hg18 chromosome 
 size data using the seqAnglPo function. The chromosome size data stored in UCSC.hg18.chr.
}
\author{
%%  ~~ possibly secondary sources and usages ~~
Ying Hu <yhu@mail.nih.gov>
Chunhua Yan <yanch@mail.nih.gov>
}



