%%
\name{UCSC.hg19.chr}
\alias{UCSC.hg19.chr}
\docType{data}
\title{
 human hg19 segment data
}
\description{
 Human hg19 chromome size and binding data obtained from UCSC Genome Browser cytogenetics table.
}
\author{
%%  ~~ possibly secondary sources and usages ~~
Ying Hu <yhu@mail.nih.gov>
Chunhua Yan <yanch@mail.nih.gov>
}



