%%
\name{UCSC.hg19}
\alias{UCSC.hg19}
\docType{data}
\title{
 human hg19 circumference coordinates
}
\description{
 Human hg19 circumference coordinates (angles) are calculated from hg19 chromosome 
 size data using the seqAnglPo function. The chromosome size data stored in UCSC.hg19.chr.
}
\author{
%%  ~~ possibly secondary sources and usages ~~
Ying Hu <yhu@mail.nih.gov>
Chunhua Yan <yanch@mail.nih.gov>
}



