%%
\name{UCSC.hg38.chr}
\alias{UCSC.hg38.chr}
\docType{data}
\title{
 human hg38 chromosome information.
}
\description{
 Human hg38 chromome size and binding data obtained from UCSC Genome Browser cytogenetics table.
}
\author{
%%  ~~ possibly secondary sources and usages ~~
Ying Hu <yhu@mail.nih.gov>
Chunhua Yan <yanch@mail.nih.gov>
}



