%%
\name{UCSC.mm10.chr}
\alias{UCSC.mm10.chr}
\docType{data}
\title{
 mouse mm10 segment data.
}
\description{
 Mouse mm10 chromsome size and binding data obtained from UCSC Genome Browser cytogenetics table.
}
\author{
%%  ~~ possibly secondary sources and usages ~~
Ying Hu <yhu@mail.nih.gov>
Chunhua Yan <yanch@mail.nih.gov>
}



