%%
\name{UCSC.mm10}
\alias{UCSC.mm10}
\docType{data}
\title{
 mouse mm10 circumference coordinates
}
\description{
 Mouse mm10 circumference coordinates (angles) are calculated from mm10 chromosome 
 size data using the seqAnglPo function. The mouse chromosome size data stored in UCSC.mm10.chr.
}
\author{
%%  ~~ possibly secondary sources and usages ~~
Ying Hu <yhu@mail.nih.gov>
Chunhua Yan <yanch@mail.nih.gov>
}



