%%
\name{UCSC.mm9.chr}
\alias{UCSC.mm9.chr}
\docType{data}
\title{
 mouse mm9 segment data.
}
\description{
 Mouse mm9 chromsome size and binding data obtained from UCSC Genome Browser cytogenetics table.
}
\author{
%%  ~~ possibly secondary sources and usages ~~
Ying Hu <yhu@mail.nih.gov>
Chunhua Yan <yanch@mail.nih.gov>
}



