%%
\name{UCSC.mm9}
\alias{UCSC.mm9}
\docType{data}
\title{
 mouse mm9 circumference coordinates
}
\description{
 Mouse mm9 circumference coordinates (angles) are calculated from mm9 chromosome 
 size data using the seqAnglPo function. The mouse chromosome size data stored in UCSC.mm9.chr.
}
\author{
%%  ~~ possibly secondary sources and usages ~~
Ying Hu <yhu@mail.nih.gov>
Chunhua Yan <yanch@mail.nih.gov>
}



